Prints a summary of a terminal node in a tree

If no arguement is provided for the parameter node, summaries are printed for every terminal node. Otherwise, the summary of just the requested node is printed.

terminalNodeSummary(tree, node = NULL)

Arguments

tree
A splinetree object

node
The number of the node that you want summarized. To see which nodes correspond to which numbers, see stPrint(tree) or treeSummary(tree). If this parameter is provided, must correspond to a valid terminal node in the tree.

Examples

split_formula <- ~HISP + WHITE + BLACK + SEX + Num_sibs + HGC_FATHER + HGC_MOTHER
tree <- splineTree(split_formula, tformula, idvar = "ID",
   data = nlsySample, degree = 1, df = 3,
   intercept = TRUE, cp = 0.005)

#> Error in formula(object): object 'tformula' not found

terminalNodeSummary(tree)

#> 
#>  node number: 8 
#>    root
#>    WHITE< 0.5
#>    HGC_MOTHER< 9.5
#>    Num_sibs< 5.5
#> 
#>  N: 61
#>  Coefficients: 21.638716730493,6.06190454077328,9.00889849472368
#> ----------
#> 
#>  node number: 9 
#>    root
#>    WHITE< 0.5
#>    HGC_MOTHER< 9.5
#>    Num_sibs>=5.5
#> 
#>  N: 36
#>  Coefficients: 22.2566488164994,7.90644930497537,9.90836805626332
#> ----------
#> 
#>  node number: 20 
#>    root
#>    WHITE< 0.5
#>    HGC_MOTHER>=9.5
#>    Num_sibs< 3.5
#>    HISP< 0.5
#> 
#>  N: 52
#>  Coefficients: 21.7830134187367,6.26350455258489,10.7774926246101
#> ----------
#> 
#>  node number: 21 
#>    root
#>    WHITE< 0.5
#>    HGC_MOTHER>=9.5
#>    Num_sibs< 3.5
#>    HISP>=0.5
#> 
#>  N: 24
#>  Coefficients: 21.7230311542365,4.59006024331612,7.02910204466869
#> ----------
#> 
#>  node number: 11 
#>    root
#>    WHITE< 0.5
#>    HGC_MOTHER>=9.5
#>    Num_sibs>=3.5
#> 
#>  N: 81
#>  Coefficients: 21.1265283315391,4.92702760984196,8.28216178535301
#> ----------
#> 
#>  node number: 6 
#>    root
#>    WHITE>=0.5
#>    SEX< 1.5
#> 
#>  N: 125
#>  Coefficients: 23.0227281240954,3.59660270648257,6.78756471063505
#> ----------
#> 
#>  node number: 7 
#>    root
#>    WHITE>=0.5
#>    SEX>=1.5
#> 
#>  N: 121
#>  Coefficients: 20.606803154073,3.66483529507145,6.63268835408635
#> ----------