Returns the tree frame.

Provides a similar output to model$frame, but with the redundant information of yval and yval2 removed. Also omits the deviance, the complexity, and the weight. Useful for viewing node numbers and for extracting coefficients for a given node.

treeSummary(model)

Arguments

model
A splinetree object.

Value

A dataframe. The number of rows is the same as the number of nodes in the tree. The row names display the node labels of each node. The "var" attribute either displays the split variable selected at each node, or <leaf> if this node is a terminal node. The "n" attribute displays the number of individuals in the node. The "dev" attribute reports the projected sum of squares at this node; terminal nodes have the smallest values for "dev" because this is what the tree building process is supposed to minimize. The "coeffs" attribute displays the coefficients predicted for each node.

Examples

nlsySubset <- nlsySample[nlsySample$ID %in% sample(unique(nlsySample$ID), 400),]
split_formula <- ~HISP + WHITE + BLACK + SEX + Num_sibs + HGC_FATHER + HGC_MOTHER
tree <- splineTree(split_formula, BMI~AGE, idvar = "ID",
   data = nlsySubset, degree = 1, df = 3,
   intercept = TRUE, cp = 0.005)

treeSummary(tree)

#>            var   n        dev  coeffs.1  coeffs.2  coeffs.3
#> 1   HGC_MOTHER 400 66922.0962 21.761225  4.674187  7.798282
#> 2   HGC_FATHER 102 24761.3721 22.421210  5.934618  8.996117
#> 4   HGC_FATHER  63 17143.6409 22.709094  6.150662  8.830353
#> 8     Num_sibs  51 11956.1155 22.379939  5.632652  8.777104
#> 16  HGC_MOTHER  39  7756.3283 21.943195  4.995032  8.729627
#> 32      <leaf>  13  1208.1379 20.784155  4.210422  8.001371
#> 33      <leaf>  26  6101.0421 22.522715  5.387337  9.093755
#> 17      <leaf>  12  3295.7425 23.799356  7.704916  8.931404
#> 9       <leaf>  12  4285.1237 24.108000  8.352207  9.056660
#> 5     Num_sibs  39  7421.2543 21.956169  5.585624  9.263889
#> 10      <leaf>  14  2052.8739 22.272771  7.607496 10.614828
#> 11      <leaf>  25  4797.9106 21.778872  4.453377  8.507363
#> 3          SEX 298 39326.1838 21.535323  4.242764  7.388285
#> 6       <leaf> 153 14165.2037 22.302053  4.408201  7.273237
#> 7        BLACK 145 23528.1360 20.726291  4.068198  7.509681
#> 14    Num_sibs 106 16320.3235 20.521299  3.558911  6.929814
#> 28    Num_sibs  96 12819.6493 20.535435  3.355437  6.291725
#> 56  HGC_MOTHER  78 11360.1570 20.900960  3.439726  6.298122
#> 112     <leaf>  66 10271.4865 21.077042  3.667877  6.793804
#> 113     <leaf>  12   561.5043 19.932504  2.184896  3.571875
#> 57      <leaf>  18   937.0544 18.951495  2.990188  6.264001
#> 29      <leaf>  10  3034.1502 20.385589  5.512257 13.055470
#> 15      <leaf>  39  6038.0258 21.283450  5.452415  9.085731