Computes a level-based evaluation metric for a splineforest that was built WITH an intercept.

Computes the R-squared metric for a splineforest object. Goal is to see how well the predicted response values match the actual response values. Note that this function should only be used on forests where the intercept parameter is TRUE. A simple 1-SSE/SST calculation.

yR2Forest(forest, method = "oob")

Arguments

forest
A splineforest object

method
How would you like to compute this metric? The choices are "oob", "itb", or "all". "oob" means that predictions for a datapoint can only be made using trees for which that datapoint was "out of the bag" (not in the bootstrap sample). "all" means that all trees are used in the prediction for every datapoint. "itb" means that predictions for a datapoint are made using only the trees for which this datapoint was IN the bootstrap sample.

Value

Returns 1-SSE/SST, where SSE is the total sum of squared errors of the true responses and predicted responses, and SST is the total sum of squared erros of the responses around their mean. If this forest was not built with an intercept, returns NULL.

Examples

yR2Forest(forest, method="all")

#> [1] 0.23554