Prints the tree frame.

Provides a similar output to model$frame, but with the redundant information of yval and yval2 removed. Also omits the deviance, the complexity, and the weight. Useful for viewing node numbers and for extracting coefficeints for a given node.

treeSummary(model)

Arguments

model
A splinetree object.

Value

A dataframe. The number of rows is the same as the number of nodes in the tree. The row names display the node labels of each node. The "var" attribute either displays the split variable selected at each node, or <leaf> if this node is a terminal node. The "n" attribute displays the number of indivduals in the node. The "dev" attribute reports the projected sum of squares at this node; terminal nodes have the smallest values for "dev" because this is what the tree building process is supposed to minimize. The "coeffs" attribute displays the coefficients predicted for each node.

Examples

split_formula <- ~HISP + WHITE + BLACK + SEX + Num_sibs + HGC_FATHER + HGC_MOTHER
tree <- splineTree(split_formula, tformula, idvar = "ID",
   data = nlsySample, degree = 1, df = 3,
   intercept = TRUE, cp = 0.005)

#> Error in formula(object): object 'tformula' not found

treeSummary(tree)

#>           var   n       dev  coeffs.1  coeffs.2  coeffs.3
#> 1       WHITE 500 82129.723 21.715567  4.764763  7.914456
#> 2  HGC_MOTHER 254 43339.776 21.600470  5.863627  9.079635
#> 4    Num_sibs  97 18837.124 21.868052  6.746478  9.342722
#> 8      <leaf>  61  7784.603 21.638717  6.061905  9.008898
#> 9      <leaf>  36 10402.857 22.256649  7.906449  9.908368
#> 5    Num_sibs 157 23583.429 21.435148  5.318171  8.917090
#> 10       HISP  76 12714.977 21.764072  5.735048  9.593790
#> 20     <leaf>  52 10073.923 21.783013  6.263505 10.777493
#> 21     <leaf>  24  2206.412 21.723031  4.590060  7.029102
#> 11     <leaf>  81 10317.506 21.126528  4.927028  8.282162
#> 3         SEX 246 36185.614 21.834407  3.630164  6.711386
#> 6      <leaf> 125 14782.485 23.022728  3.596603  6.787565
#> 7      <leaf> 121 18931.142 20.606803  3.664835  6.632688